requirements.txt
openai
python-dotenv
pinecone
pyTelegramBotAPI
haystack-ai
pinecone-haystack
requests
docstring-parser
jsonschema